requirements.txt
Pillow
pyscreenshot
argparse
pyinstaller
PyQt5
